+++ concordile-broker/src/main/java/io/github/concordile/broker/mapper/v1/VerificationResponseMapper.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2025-present The Concordile Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.github.concordile.broker.mapper.v1;
+
+import io.github.concordile.broker.api.v1.VerificationResponse;
+import io.github.concordile.broker.config.MapStructConfig;
+import io.github.concordile.broker.domain.Verification;
+import io.github.concordile.broker.mapper.ResponseMapper;
+import org.mapstruct.Mapper;
+
+@Mapper(config = MapStructConfig.class, uses = VerificationResultResponseMapper.class)
+public interface VerificationResponseMapper
+        extends ResponseMapper<Verification, VerificationResponse> {
+
+    @Override
+    VerificationResponse mapDomain2Response(Verification domain);
+
+}
+
